Splitting classes to separate files from DataRouter.
