+++ src/dotnet-counters-ui/dotnet-counters-ui/DataRouter.cs
@@ -109,26 +109,4 @@ namespace DotnetCountersUi
             }
         };
     }
-
-    public class GraphData
-    {
-        public string Name { get; set; }
-    }
-
-    public interface IDataRouter
-    {
-        void Start(int pid);
-        GraphData Register(string graphId, Action<double> action);
-    }
-
-    public class SeriesMetadata
-    {
-        public string Id { get; init; }
-        public string Name { get; init; }
-        public string? StackGroup { get; init; }
-        public List<object> X { get; init; } = new List<object>();
-        public List<object> Y { get; init; } = new List<object>();
-        public Func<IDictionary<string, object>, bool> Filter { get; init; }
-        public Func<IDictionary<string,object>, double> Selector { get; init; }
-    }
 }
