+++ src/dotnet-counters-ui/dotnet-counters-ui/GraphData.cs
@@ -1,0 +1,6 @@
+namespace DotnetCountersUi;
+
+public class GraphData
+{
+  public string Name { get; set; }
+}
