+++ src/dotnet-counters-ui/dotnet-counters-ui/IDataRouter.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace DotnetCountersUi;
+
+public interface IDataRouter
+{
+  void Start(int pid);
+  GraphData Register(string graphId, Action<double> action);
+}
